// src/ast-types.ts
export namespace ExprNS {
    export interface Visitor<T> {
        visitBigIntLiteralExpr(expr: BigIntLiteral): T
        visitBinaryExpr(expr: Binary): T
        visitCompareExpr(expr: Compare): T
        visitBoolOpExpr(expr: BoolOp): T
        visitGroupingExpr(expr: Grouping): T
        visitLiteralExpr(expr: Literal): T
        visitUnaryExpr(expr: Unary): T
        visitTernaryExpr(expr: Ternary): T
        visitLambdaExpr(expr: Lambda): T
        visitMultiLambdaExpr(expr: MultiLambda): T
        visitVariableExpr(expr: Variable): T
        visitCallExpr(expr: Call): T
        visitComplexExpr(expr: Complex): T
        visitNoneExpr(expr: None): T
    }
    export abstract class Expr {
        startToken: Token;
        endToken: Token;
        protected constructor(startToken: Token, endToken: Token) {
            this.startToken = startToken;
            this.endToken = endToken;
        }
        abstract accept(visitor: Visitor<any>): any;
    }
    export class None extends Expr { 
        constructor(startToken: Token, endToken: Token, value: string = "None") {
            super(startToken, endToken);
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitNoneExpr(this);
        }
    }
    export class BigIntLiteral extends Expr {
        value: string;
        constructor(startToken: Token, endToken: Token, value: string){
            super(startToken, endToken)
            this.value = value;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitBigIntLiteralExpr(this)
        }
    }
    export class Complex extends Expr {
        value: PyComplexNumber;
        constructor(startToken: Token, endToken: Token, value: string) {
            super(startToken, endToken);
            this.value = PyComplexNumber.fromString(value);;
        }
        override accept<T>(visitor: Visitor<T>): T {
            return visitor.visitComplexExpr(this);
        }
    }
    export class Binary extends Expr {
        left: Expr;
        operator: Token;
        right: Expr;
        constructor(startToken: Token, endToken: Token, left: Expr, operator: Token, right: Expr){
            super(startToken, endToken)
            this.left = left;
            this.operator = operator;
            this.right = right;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitBinaryExpr(this)
        }
    }
    export class Compare extends Expr {
        left: Expr;
        operator: Token;
        right: Expr;
        constructor(startToken: Token, endToken: Token, left: Expr, operator: Token, right: Expr){
            super(startToken, endToken)
            this.left = left;
            this.operator = operator;
            this.right = right;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitCompareExpr(this)
        }
    }
    export class BoolOp extends Expr {
        left: Expr;
        operator: Token;
        right: Expr;
        constructor(startToken: Token, endToken: Token, left: Expr, operator: Token, right: Expr){
            super(startToken, endToken)
            this.left = left;
            this.operator = operator;
            this.right = right;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitBoolOpExpr(this)
        }
    }
    export class Grouping extends Expr {
        expression: Expr;
        constructor(startToken: Token, endToken: Token, expression: Expr){
            super(startToken, endToken)
            this.expression = expression;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitGroupingExpr(this)
        }
    }
    export class Literal extends Expr {
        value: true | false | number | string;
        constructor(startToken: Token, endToken: Token, value: true | false | number | string){
            super(startToken, endToken)
            this.value = value;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitLiteralExpr(this)
        }
    }
    export class Unary extends Expr {
        operator: Token;
        right: Expr;
        constructor(startToken: Token, endToken: Token, operator: Token, right: Expr){
            super(startToken, endToken)
            this.operator = operator;
            this.right = right;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitUnaryExpr(this)
        }
    }
    export class Ternary extends Expr {
        predicate: Expr;
        consequent: Expr;
        alternative: Expr;
        constructor(startToken: Token, endToken: Token, predicate: Expr, consequent: Expr, alternative: Expr){
            super(startToken, endToken)
            this.predicate = predicate;
            this.consequent = consequent;
            this.alternative = alternative;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitTernaryExpr(this)
        }
    }
    export class Lambda extends Expr {
        parameters: Token[];
        body: Expr;
        constructor(startToken: Token, endToken: Token, parameters: Token[], body: Expr){
            super(startToken, endToken)
            this.parameters = parameters;
            this.body = body;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitLambdaExpr(this)
        }
    }
// ...
    export class Variable extends Expr {
        name: Token;
        constructor(startToken: Token, endToken: Token, name: Token){
            super(startToken, endToken)
            this.name = name;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitVariableExpr(this)
        }
    }
    export class Call extends Expr {
        callee: Expr;
        args: Expr[];
        constructor(startToken: Token, endToken: Token, callee: Expr, args: Expr[]){
            super(startToken, endToken)
            this.callee = callee;
            this.args = args;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitCallExpr(this)
        }
    }
    
    
}
export class ImportAlias {
    constructor(public name: Token, public alias: Token | null) {}
}

export namespace StmtNS {
// ...
    export abstract class Stmt {
        startToken: Token;
        endToken: Token;
        protected constructor(startToken: Token, endToken: Token) {
            this.startToken = startToken;
            this.endToken = endToken;
        }
        abstract accept(visitor: Visitor<any>): any;
    }
// ...
    export class Assign extends Stmt {
        name: Token;
        value: ExprNS.Expr;
        constructor(startToken: Token, endToken: Token, name: Token, value: ExprNS.Expr){
            super(startToken, endToken)
            this.name = name;
            this.value = value;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitAssignStmt(this)
        }
    }
// ...
    export class Return extends Stmt {
        value: ExprNS.Expr | null;
        constructor(startToken: Token, endToken: Token, value: ExprNS.Expr | null){
            super(startToken, endToken)
            this.value = value;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitReturnStmt(this)
        }
    }
// ...
    export class If extends Stmt {
        condition: ExprNS.Expr;
        body: Stmt[];
        elseBlock: Stmt[] | null;
        constructor(startToken: Token, endToken: Token, condition: ExprNS.Expr, body: Stmt[], elseBlock: Stmt[] | null){
            super(startToken, endToken)
            this.condition = condition;
            this.body = body;
            this.elseBlock = elseBlock;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitIfStmt(this)
        }
    }
// ...
    export class FunctionDef extends Stmt {
        name: Token;
        parameters: Token[];
        body: Stmt[];
        varDecls: Token[];
        constructor(startToken: Token, endToken: Token, name: Token, parameters: Token[], body: Stmt[], varDecls: Token[]){
            super(startToken, endToken)
            this.name = name;
            this.parameters = parameters;
            this.body = body;
            this.varDecls = varDecls;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitFunctionDefStmt(this)
        }
    }
    export class SimpleExpr extends Stmt {
        expression: ExprNS.Expr;
        constructor(startToken: Token, endToken: Token, expression: ExprNS.Expr){
            super(startToken, endToken)
            this.expression = expression;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitSimpleExprStmt(this)
        }
    }
    export class FileInput extends Stmt {
        statements: Stmt[];
        varDecls: Token[];
        constructor(startToken: Token, endToken: Token, statements: Stmt[], varDecls: Token[]){
            super(startToken, endToken)
            this.statements = statements;
            this.varDecls = varDecls;
        }
        override accept(visitor: Visitor<any>): any {
            return visitor.visitFileInputStmt(this)
        }
    }
    
    
}